Add constant derivative BC and move GenericBCs to a new header
